+++ app/Providers/AppServiceProvider.php
@@ -58,6 +58,7 @@ class AppServiceProvider extends ServiceProvider
             $service->registerHandler($app->make(\App\Services\ChoiceHandlers\SpellChoiceHandler::class));
             $service->registerHandler($app->make(\App\Services\ChoiceHandlers\HitPointRollChoiceHandler::class));
             $service->registerHandler($app->make(\App\Services\ChoiceHandlers\AbilityScoreChoiceHandler::class));
+            $service->registerHandler($app->make(\App\Services\ChoiceHandlers\FeatChoiceHandler::class));
 
             return $service;
         });